twl6030: Change channels to accomodate channel 16
TWL6030 has total of 17 channels from 0 to 16. Currently the
"channels" variable of "twl6030_gpadc_request" structure is u16.

Changing it to u32 so that it can accomodate channel 16.

+++ drivers/mfd/twl6030-gpadc.c
@@ -272,7 +272,7 @@ static int twl6030_gpadc_channel_raw_read(struct twl6030_gpadc_data *gpadc,
 }
 
 static int twl6030_gpadc_read_channels(struct twl6030_gpadc_data *gpadc,
-		u8 reg_base, u16 channels, struct twl6030_gpadc_request *req)
+		u8 reg_base, u32 channels, struct twl6030_gpadc_request *req)
 {
 	int count = 0;
 	u8 reg, i;
